refactor(work): extract first_shipped helpers from sqlite_service to backend-neutral module (#1737)
Move the first-shipped milestone helpers (``mark_first_shipped``,
``_record_first_shipped_activity``, ``task_landed_commit``,
``maybe_record_first_shipped``) plus their backend-neutral support
(``state_path`` / ``load_state`` / ``first_shipped_at`` /
``_HasExecutions`` Protocol / ``_STATE_FILENAME``) out of
``sqlite_service.py`` into a new leaf module
``pollypm.work.first_shipped``.

``pg_service.py`` now imports ``maybe_record_first_shipped`` from
the new leaf module instead of reaching into the sqlite-only
service, removing the last runtime dependency that blocked
pg_service.py from being treated as production code. The eventual
K-source-ripout can now delete ``sqlite_service.py`` atomically.

``sqlite_service.py`` re-exports the helpers under the original
names so existing in-tree callers (including the
``maybe_record_first_shipped`` monkeypatch in
``tests/test_pg_gap_sweep_2.py``) keep working unchanged. Pure
code move — function signatures and behavior are identical.
